+++ noodles-cram/src/record/resolve.rs
@@ -57,6 +57,12 @@ mod internal {
             }
 
             match feature {
+                Feature::Bases(_, bases) => {
+                    for &base in bases {
+                        buf[read_pos] = base;
+                        read_pos += 1;
+                    }
+                }
                 Feature::Substitution(_, code) => {
                     let base = raw_reference_sequence[ref_pos] as char;
                     let reference_base = Base::try_from(base).unwrap_or_default();
@@ -167,6 +173,7 @@ mod tests {
         };
 
         t(&[], b"ACGT");
+        t(&[Feature::Bases(1, b"TGCA".to_vec())], b"TGCA");
         t(&[Feature::Substitution(2, 1)], b"AGGT");
         t(&[Feature::Insertion(2, b"GG".to_vec())], b"AGGC");
         t(&[Feature::Deletion(2, 2)], b"ATAC");